Add (empty) default values for various form fields
This changes the recurring contribution form to add an empty default
value for some fields, and for fields where that's appropriate, an
actual default value.

The goal is to trigger the required fields validation when users skip
over a field rather than selecting the value that just happens to be
the first option in the <select>.

This is mostly useful in connection with rcont_remember_values set to
"0", as the form would otherwise just prefill the previously entered
values.